// packages/backend/src/question/models/question-type.enum.ts
export enum EQuestionType {
  MULTIPLE_CHOICE = 'MULTIPLE_CHOICE',
  TRUE_FALSE = 'TRUE_FALSE',
  FILL_IN_THE_BLANK = 'FILL_IN_THE_BLANK',
  ORDERING = 'ORDERING',
  ESTIMATE = 'ESTIMATE',
  NUMERIC = 'NUMERIC',
  IMAGE_QUESTION = 'IMAGE_QUESTION',
  MUSIC_QUESTION = 'MUSIC_QUESTION',
  VIDEO_QUESTION = 'VIDEO_QUESTION',
}